pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workout-reminder-bot"
version = "1.0.0"
description = "Telegram bot for workout reminders"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "aiogram>=3.0.0,<4.0.0",
    "apscheduler>=3.10.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytz>=2023.3",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "pytest",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.setuptools.packages.find]
where = ["src"]
